User commit

+++ src/main/java/com/example/demo/service/impl/ShipmentServiceImpl.java
@@ -9,24 +9,33 @@ import com.example.demo.repository.ShipmentRepository;
 import com.example.demo.repository.VehicleRepository;
 import com.example.demo.service.ShipmentService;
 import org.springframework.stereotype.Service;
+
 import java.time.LocalDate;
+
 @Service
 public class ShipmentServiceImpl implements ShipmentService {
+
     private final ShipmentRepository shipmentRepository;
     private final VehicleRepository vehicleRepository;
     private final LocationRepository locationRepository;
-    public ShipmentServiceImpl(ShipmentRepository shipmentRepository,
-                               VehicleRepository vehicleRepository,
-                               LocationRepository locationRepository) {
+
+    public ShipmentServiceImpl(
+            ShipmentRepository shipmentRepository,
+            VehicleRepository vehicleRepository,
+            LocationRepository locationRepository) {
+
         this.shipmentRepository = shipmentRepository;
         this.vehicleRepository = vehicleRepository;
         this.locationRepository = locationRepository;
     }
+
     @Override
     public Shipment createShipment(Long vehicleId, Shipment shipment) {
+
         Vehicle vehicle = vehicleRepository.findById(vehicleId)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Vehicle not found"));
+
         Location pickup = locationRepository.findById(
                 shipment.getPickupLocation().getId())
                 .orElseThrow(() ->
@@ -36,20 +45,31 @@ public class ShipmentServiceImpl implements ShipmentService {
                 shipment.getDropLocation().getId())
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Location not found"));
-        if (shipment.getWeightKg() > vehicle.getCapacityKg()) {
-            throw new IllegalArgumentException("Weight exceeds vehicle capacity");
+
+        // Validate weight
+        if (shipment.getWeightKg().doubleValue() >
+                vehicle.getCapacityKg().doubleValue()) {
+
+            throw new IllegalArgumentException(
+                    "Weight exceeds vehicle capacity");
         }
+
+        // Validate date
         if (shipment.getScheduledDate().isBefore(LocalDate.now())) {
-            throw new IllegalArgumentException("Date cannot be in the past");
+            throw new IllegalArgumentException(
+                    "Date cannot be in the past");
         }
+
         shipment.setVehicle(vehicle);
         shipment.setPickupLocation(pickup);
         shipment.setDropLocation(drop);
 
         return shipmentRepository.save(shipment);
     }
+
     @Override
     public Shipment getShipment(Long shipmentId) {
+
         return shipmentRepository.findById(shipmentId)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Shipment not found"));
